Removed memory leak from modules
Spoiler: geoip-lite was the issue
