+++ cluster.js
@@ -2,7 +2,7 @@ let cluster = require('cluster');
 let numCPUs = require('os').cpus().length;
 
 if (cluster.isMaster) {
-  for (let i = 0; i < numCPUs/2; i++) {
+  for (let i = 0; i < numCPUs; i++) {
     cluster.fork();
   }
   cluster.on('exit', function(worker, code, signal) {
