+++ remastered_modules/params/acceptHeaders.js
@@ -1,0 +1,8 @@
+module.exports = function(next) {
+  next(null, {
+    acceptHeaders:{
+      accept:this.req.headers['accept'],
+      language:this.req.headers['accept-language']
+    }
+  })
+}
