+++ remastered_modules/params/geoip.js
@@ -1,0 +1,8 @@
+module.exports = function(next) {
+  const geo = "Ukraine"
+  next(null, {
+    geoip: {
+      country: geo ? geo.country : null
+    }
+  })
+}
