+++ remastered_modules/params/index.js
@@ -1,0 +1,9 @@
+let useragent = require('./useragent.js');
+let acceptHeaders = require('./acceptHeaders.js');
+let geoip = require('./geoip.js');
+
+module.exports = {
+  useragent,
+  acceptHeaders,
+  geoip
+}
